update: refactor landing view

+++ routes/web.php
@@ -1,23 +1,22 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\Landing;
+use App\Http\Controllers\LandingController;
 use App\Http\Controllers\VacancyController;
 use App\Http\Controllers\Admin\AuthController;
 
-Route::get("/", [Landing::class, "index"]);
-Route::get("/vacancies", [Landing::class, "explore"]);
-Route::get("/vacancy/{id}", [Landing::class, "detail"]);
+Route::get("/", [LandingController::class, "index"])->name("home");
+Route::get("/vacancies", [LandingController::class, "explore"])->name("vacancies");
+Route::get("/vacancy/{id}", [LandingController::class, "detail"])->name("vacancy.detail");
 
 Route::prefix("admin")->group(function () {
     Route::middleware("admin.guest")->group(function () {
-        Route::get("/", [AuthController::class, "showLoginForm"]);
         Route::get("/login", [AuthController::class, "showLoginForm"])->name("admin.login");
         Route::post("/login", [AuthController::class, "login"])->name("admin.login.post");
     });
 
     Route::middleware("admin.auth")->group(function () {
-        Route::get("/logout", [AuthController::class, "logout"])->name("admin.logout");
+        Route::post("/logout", [AuthController::class, "logout"])->name("admin.logout");
 
         Route::prefix("vacancy")->group(function () {
             Route::get("/", [VacancyController::class, "index"])->name("admin.vacancies");
@@ -26,8 +25,11 @@ Route::prefix("admin")->group(function () {
             Route::get("/edit/{id}", [VacancyController::class, "edit"])->name("admin.vacancy.edit");
             Route::post("insert", [VacancyController::class, "store"]);
             Route::put("/update/{id}", [VacancyController::class, "update"])->name("admin.vacancy.update");
-            Route::post("/{id}/thumbnail", [VacancyController::class, "updateThumbnail"]);
-            Route::delete("/delete/{id}", [VacancyController::class, "destroy"]);
+            // Route::post("upload-thumbnail/temp-store", [VacancyController::class, "storeTempThumbnail"])->name("admin.vacancy.upload-thumbnail.temp-store");
+            // Route::post("upload-thumbnail/temp-delete", [VacancyController::class, "deleteTempThumbnail"])->name("admin.vacancy.upload-thumbnail.temp-delete");
+            Route::post("/upload-thumbnail/{id}", [VacancyController::class, "updateThumbnail"])->name("admin.vacancy.upload-thumbnail");
+            Route::delete("/delete/{id}", [VacancyController::class, "destroy"])->name("admin.vacancy.delete");
+            Route::post("/change-status/{id}", [VacancyController::class, "changeStatus"])->name("admin.vacancy.change-status");
         });
     });
 });
